Cleanup and Testing (#28)
* Cleanup and Testing

- Added .vscode to the .gitignore and removed it from the project.
- Added SECRET_KEY to application config and set testing for development
config to true.
- Removed unneeded imports from Authentication.py and updated
capitalization of API to match other classes and packages.
- Moved email based id listing left over from old system to integer
based id listing.
- Made methods of API static as they do not need access to an API
instance and fixed queries to use proper ids.
- Updated DBModels to use regular integer primary keys, which triggers
SQLite to use a special unique table row system as described at:
http://docs.sqlalchemy.org/en/latest/dialects/sqlite.html. Also updated
naming scheme to match our site terminology (moved lecture to class as
lectures will be used to store groups of questions) and changed table
names to be plural (a table holds a collection of 'Users' for example,
but the model operates on single 'User' objects).
- Added additional styling to our hosted content, rather than CDN, so
that we have full access to the unminifed versions for debugging as well
as being able to run the full application entirely offline.
- Updated naming of student lectures in the Angular controller to
reflect the change from lecture->class. Also fixed response.data for
set_student_classes checking the wrong type and not successfully adding
a class as a result.
- Updated layout with script and style resources for self hosting.
- Updated class list view to use class list controller instead of
lectures controller.
- Added database test content to app initialization. If TESTING is set
to true for the current application content, then the database contents
will be DROPPED and recreated and filled with fresh test data. This
should allow us to have a fully functioning app with reasonable test
data for the start of every demo.

* Moved API from api.py to API.py

* Adding Library Files

Lib was in the gitignore so the new angular offline files were not added
properly

* Updated Return Type

Updated the return type of create_prof_class to us json.dumps