Lecture Ending (#67)
* Initial Work

Initial work on the live lecture stop button

* Final Professor Work

Added the end lecture button and API method for the professor

* Student Lecture Polling

Added live lecture polling for the student so that when the professor
ends the lecture, the student is returned to the lecture list

* Lecture Title

Added the lecture title to the live lecture view

* Polling Question Submit

Updated polling question submission to use a submit button since we talk
about that as a lesson learned from usability testing

* Update Gauge Reset

Gauge reset now uses the live lecture id instead of the lecture id

* Invalid Lecture Redirect

Redirect to class list if the given live lecture does not exist

* Fixed Go Live

Go live was redirecting before the new live id was set, causing it to
always redirect to live lecture with id 1. Also changed it to use the
class id, which is what get live lecture uses. Updated logout button
style so that it looks nice.

* Updated Class List Styling

New style for the class list

* More Styling

Updated styling on other pages to match a consistent theme

* Updating liveView.html

* Cleanup Polling Question

* Readded db

* DB.