Merging Sprint 5 into master (#69)
* Base DB Models

Added the basic database model structure

* Data Initialization Update

Updated __init__.py to use the new database schema. Added the new
database with initial data. Small fixes to DBModels.py.

* Gitignore Update

Added the SQLite temporary journal file to the .gitignore

* Foreign Key Cycle Resolution

Updated DB schema to resolve foreign key cycle

* API Updates

Updated the API to use the new live lecture stuff

* Controller Updates

Updated controllers to follow new DB arch and lecture flow

* Controller Testing

Readded controllers and some testing

* Lecture Fix

Fixing lecture name and class id being backwards when creating a new
lecture

* Professor Check Update

Updated is_professor() in API to use boolean not instead of bitwise not

* API and controllers (#56)

* api and controllers

* fixed typo

* fixed errors

* Just integrated google sign in. Sign out doesn't work yet

* Modified class List page

* Fixed error

* Neilv sprint4 (#59)

* flask student auth works, js does not

* Auth gets id, needs more robust error checking

* implemented lecture controllers

* fixed add lecture and student functionality

* gauge handling updated

* updated gauge methods pt 2

* use string handling for negative URL numbers

* DB Architecture and Clean Up (#55)

* Base DB Models

Added the basic database model structure

* Data Initialization Update

Updated __init__.py to use the new database schema. Added the new
database with initial data. Small fixes to DBModels.py.

* Gitignore Update

Added the SQLite temporary journal file to the .gitignore

* Foreign Key Cycle Resolution

Updated DB schema to resolve foreign key cycle

* API Updates

Updated the API to use the new live lecture stuff

* Controller Updates

Updated controllers to follow new DB arch and lecture flow

* Controller Testing

Readded controllers and some testing

* Lecture Fix

Fixing lecture name and class id being backwards when creating a new
lecture

* Professor Check Update

Updated is_professor() in API to use boolean not instead of bitwise not

* missed an error

* what is happening?

* didn't do git add

* implemented lecture list

* Implemented screens for viewing and adding polling questions for a lecture.

* new update method, removed routparams

* removed debug print statement

* Fixing changes commented on pull request.

* Extending baseHomepage.html in HTML files now.

* Fix merge conflict and have username show up in PollingQuestionsList.html.

* Display username on lecture list page.

* Controllers and Methods for Polling Questions (#63)

* Professor Polling Question UI (#64)

* Polling Question Listing

Added the ability to list questions in the professor live view. Updated
the controllers to be able to handle this.

* Professor Question Controll

Added the ability for professors to start and stop presenting questions,
as well as view the results of presenting questions.

* Added LiveViewController Changes

Commit of LiveViewController.js that got missed in the last commit

* Fixed Polling Question Creation

Updated polling question creation to send the answer choice as 'A', 'B',
'C', or 'D' per the design

* Stop Polling Question

Updated the API to clear out the current polling question id when
stopping a polling question

* Cleanup. (#66)

* Cleanup.

* Removing debug statements.

* Student UI for Polling Questions (#65)

* Polling Question Listing

Added the ability to list questions in the professor live view. Updated
the controllers to be able to handle this.

* Professor Question Controll

Added the ability for professors to start and stop presenting questions,
as well as view the results of presenting questions.

* Added LiveViewController Changes

Commit of LiveViewController.js that got missed in the last commit

* Fixed Polling Question Creation

Updated polling question creation to send the answer choice as 'A', 'B',
'C', or 'D' per the design

* student ui for polling questions

* fixed bug

* Parameter Fix

Added missing parameters to respond_to_polling_question() in the Flask
lecture controller

* Flip booleans on which buttons display on classList.html

* Fixed Auth bug

* Lecture Ending (#67)

* Initial Work

Initial work on the live lecture stop button

* Final Professor Work

Added the end lecture button and API method for the professor

* Student Lecture Polling

Added live lecture polling for the student so that when the professor
ends the lecture, the student is returned to the lecture list

* Lecture Title

Added the lecture title to the live lecture view

* Polling Question Submit

Updated polling question submission to use a submit button since we talk
about that as a lesson learned from usability testing

* Update Gauge Reset

Gauge reset now uses the live lecture id instead of the lecture id

* Invalid Lecture Redirect

Redirect to class list if the given live lecture does not exist

* Fixed Go Live

Go live was redirecting before the new live id was set, causing it to
always redirect to live lecture with id 1. Also changed it to use the
class id, which is what get live lecture uses. Updated logout button
style so that it looks nice.

* Updated Class List Styling

New style for the class list

* More Styling

Updated styling on other pages to match a consistent theme

* Updating liveView.html

* Cleanup Polling Question

* Readded db

* DB.

* Adding Procfile

* Changed from txt

* Updating init.py

* Procfile thing